refactor: removed GameStatistics class, and incorporated into StatTracker
