+++ runner.rb
@@ -12,3 +12,5 @@ season_data = {
 }
 
 stat_tracker = StatTracker.from_csv(season_data)
+puts stat_tracker.highest_total_score
+puts stat_tracker.class
