+++ spec/spec_helper.rb
@@ -9,4 +9,4 @@ require './lib/team'
 require './lib/season'
 require './lib/stat_tracker'
 require 'csv'
-require './lib/game_statistics'
+
